+++ src/App.jsx
@@ -1,37 +1,23 @@
-import React, { useState } from 'react';
-import reactLogo from './assets/react.svg';
-import viteLogo from '../public/vite.svg';
-import './App.css';
+import React from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import './color.css';
+import MainPage from './pages/MainPage';
+import SendMessagePage from './pages/SendMessagePage';
+import CreatedRollingPage from './pages/CreatedRollingPage';
+import CreatedRolloingPageEdit from './pages/CreatedRollingPageEdit';
+import PostPage from './pages/MakePostPage';
 
 function App() {
-  const [count, setCount] = useState(0);
-
   return (
-    <>
-      <div>
-        <a href="https://vitejs.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button type="button" onClick={() => setCount((c) => c + 1)}>
-          count is
-          {count}
-        </button>
-        <p>
-          Edit
-          <code>src/App.jsx</code>
-          and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-    </>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<MainPage />} />
+        <Route path="post/:id" element={<CreatedRollingPage />} />
+        <Route path="post/:id/edit" element={<CreatedRolloingPageEdit />} />
+        <Route path="post/:id/message" element={<SendMessagePage />} />
+        <Route path="post" elemet={<PostPage />} />
+      </Routes>
+    </BrowserRouter>
   );
 }
